Use OpenGL 3.2 Core Profile on OS X
Yet another effort to pointlessly make the app comply with some
arbitrary standards that don’t add any new features, better performance
or anything, but are „the right thing to do“. What can I say, I was
bored.

The existing code already complies with OpenGL 3.2 core profile, with
one exception: The shader is in GLSL 1.20, not 1.50. The OpenGL
implementation on OS X rejects shaders with the older version when
running in Core Profile, so there is a lot of include and #define
hackery happening to make the code work on 1.50 without actually
changing anything besides the header lines.

(Changing the shader to be 1.50 in full would have been easy, too, but
I have no idea what effects this might have on other platforms.)