Add Joystick/Gamepad support for OS X
Known issues:
- Triggers don’t work, because the mapping between them on a gamepad
and whatever OS X ends up reporting is not obvious.
- Also modern „Made for iPhone“ gamepads may or may not work; I’ll test
that next. This code uses the good old HID Manager APIs. I’m keeping
this either way because I do have a joystick that works only with them.
- The mapping of keys and axis is somewhat imperfect especially on
joysticks, because it was designed for gamepad first. I didn’t want to
change the original behavior.
- This code implements its own dead zone detection, with a hardcoded
dead zone of 10% from the center. Should this be part of the
platform-independent code instead?